+++ spec/eve_online/esi/universe_factions_spec.rb
@@ -72,6 +72,14 @@ describe EveOnline::ESI::UniverseFactions do
     end
   end
 
+  describe '#etag' do
+    let(:current_etag) { double }
+
+    before { expect(subject).to receive(:current_etag).and_return(current_etag) }
+
+    specify { expect(subject.etag).to eq(current_etag) }
+  end
+
   describe '#scope' do
     specify { expect(subject.scope).to eq(nil) }
   end